Revert "bugfixing, update template and mod.lua"
This reverts commit a1f63fd6033f5192566d8ff418f208eadf789fb6.
