--- mod_template/scripts/main.lua
@@ -1,3 +1,0 @@
-function Mod:init()
-    print("Loaded "..self.info.name.."!")
-end
